refactor(pci-project): remove contacts & right useless files
ref: #MANAGER-19780
