--- packages/manager/apps/pci-project/public/translations/contacts/Messages_de_DE.json
@@ -1,26 +1,0 @@
-{
-  "cpb_rights_expl2": "Der Admin-Kontakt kann auch Kontakte mit \"Lesen/Schreiben\" und \"Nur Lesen\" Berechtigungen zu diesem Projekt hinzufügen.",
-  "cpb_rights_owner": "Admin-Kontakt:",
-  "cpb_rights_billing_contact": "Rechnungs-Kontakt:",
-  "cpb_rights_modify": "Ändern",
-  "cpb_rights_add": "Hinzufügen",
-  "cpb_rights_table_rights": "Rechte",
-  "cpb_rights_error": "Die Informationen zu den Berechtigungen konnten nicht abgerufen werden.",
-  "cpb_rights_add_error": "Die Berechtigung konnte nicht hinzugefügt werden.",
-  "cpb_rights_remove_error": "Die Berechtigung konnte nicht gelöscht werden.",
-  "cpb_rights_table_rights_value_readOnly": "Nur Lesen",
-  "cpb_rights_table_rights_value_readWrite": "Lesen/Schreiben",
-  "cpb_rights_table_rights_add_success": "Die Berechtigung wurde erfolgreich hinzugefügt.",
-  "cpb_rights_table_rights_remove_success": "Die Berechtigung wurde erfolgreich gelöscht.",
-  "cpb_rights_add_title": "Kontakt hinzufügen",
-  "cpb_rights_delete_title": "Einen Kontakt löschen",
-  "cpb_rights_delete_question": "Sind Sie sicher, dass Sie die Rechte des Benutzers {{nickname}} löschen möchten?",
-  "common_add": "Hinzufügen",
-  "delete": "Löschen",
-  "common_cancel": "Abbrechen",
-  "common_confirm": "Bestätigen",
-  "cpb_rights_note": "Hinweis: Kontakte mit \"Lesen/Schreiben\" Berechtigungen verfügen über Rechte (Änderungen an der Infrastruktur, Verwendung des Storage), die zusätzliche Kosten verursachen können. Kontakte mit \"Nur Lesen\" Berechtigungen können keine Aktionen für das Projekt durchführen. Die Standardzahlungsart des Rechnungs-Kontakts wird für alle diese Rechnungen verwendet.",
-  "cpb_rights_table_nichandle": "Kundenkennung ",
-  "cpb_rights_table_email": "E-Mail",
-  "pci_projects_project_contacts_title": "Kontakt &amp; Rechte"
-}
