--- packages/manager/apps/pci-project/public/translations/contacts/Messages_en_GB.json
@@ -1,26 +1,0 @@
-{
-  "cpb_rights_expl2": "The admin contact can also add contacts with read-write access or read-only access for this project.",
-  "cpb_rights_owner": "Admin contact:",
-  "cpb_rights_billing_contact": "Billing contact:",
-  "cpb_rights_modify": "Edit",
-  "cpb_rights_add": "Add",
-  "cpb_rights_table_rights": "Rights",
-  "cpb_rights_error": "Oops! Unable to retrieve data on the rights.",
-  "cpb_rights_add_error": "Oups! Unable to add the right.",
-  "cpb_rights_remove_error": "Oops! Unable to delete the right.",
-  "cpb_rights_table_rights_value_readOnly": "Read-only",
-  "cpb_rights_table_rights_value_readWrite": "Read/Write",
-  "cpb_rights_table_rights_add_success": "The right has been added.",
-  "cpb_rights_table_rights_remove_success": "The right has been deleted.",
-  "cpb_rights_add_title": "Add Contact",
-  "cpb_rights_delete_title": "Delete a contact",
-  "cpb_rights_delete_question": "Are you sure you want to revoke the rights for {{nickname}}?",
-  "common_add": "Add",
-  "delete": "Delete",
-  "common_cancel": "Cancel",
-  "common_confirm": "Confirm",
-  "cpb_rights_note": "NB: Contacts with read/write access have certain permissions (infrastructure updates, access to storage) that may incur an additional cost. Read-only contacts cannot make any changes to the project. The billing contact's default method of payment is used for the payment of all bills.",
-  "cpb_rights_table_nichandle": "Customer ID",
-  "cpb_rights_table_email": "Email address",
-  "pci_projects_project_contacts_title": "Contacts &amp; Rights"
-}
