--- packages/manager/apps/pci-project/public/translations/contacts/Messages_es_ES.json
@@ -1,26 +1,0 @@
-{
-  "cpb_rights_expl2": "El contacto administrador también puede añadir a este proyecto contactos en modo lectura-escritura o solo lectura.",
-  "cpb_rights_owner": "Contacto administrador: ",
-  "cpb_rights_billing_contact": "Contacto de facturación:",
-  "cpb_rights_modify": "Editar",
-  "cpb_rights_add": "Añadir",
-  "cpb_rights_table_rights": "Permisos",
-  "cpb_rights_error": "¡Vaya! No se ha podido cargar la información sobre los permisos.",
-  "cpb_rights_add_error": "¡Vaya! No se ha podido añadir el permiso.",
-  "cpb_rights_remove_error": "¡Vaya! No se ha podido eliminar el permiso.",
-  "cpb_rights_table_rights_value_readOnly": "Solo lectura",
-  "cpb_rights_table_rights_value_readWrite": "Lectura-Escritura",
-  "cpb_rights_table_rights_add_success": "El permiso se ha añadido correctamente.",
-  "cpb_rights_table_rights_remove_success": "El permiso se ha eliminado correctamente.",
-  "cpb_rights_add_title": "Añadir un contacto",
-  "cpb_rights_delete_title": "Eliminar un contacto",
-  "cpb_rights_delete_question": "¿Seguro que desea eliminar los permisos de {{nickname}}?",
-  "common_add": "Añadir",
-  "delete": "Eliminar",
-  "common_cancel": "Cancelar",
-  "common_confirm": "Confirmar",
-  "cpb_rights_note": "Importante: Los contactos en modo lectura-escritura tienen permisos (modificaciones de la infraestructura, uso del almacenamiento) que pueden generar costes adicionales. Los contactos en modo de solo lectura no pueden realizar ninguna acción en el proyecto. Para cada factura se utilizará la forma de pago por defecto del contacto de facturación.",
-  "cpb_rights_table_nichandle": "Identificador de cliente",
-  "cpb_rights_table_email": "Correo electrónico",
-  "pci_projects_project_contacts_title": "Contactos y permisos"
-}
