--- packages/manager/apps/pci-project/public/translations/contacts/Messages_fr_FR.json
@@ -1,26 +1,0 @@
-{
-  "cpb_rights_expl2": "Le contact admin peut aussi ajouter des contacts en lecture/écriture ou en lecture seule pour ce projet.",
-  "cpb_rights_owner": "Contact admin :",
-  "cpb_rights_billing_contact": "Contact facturation :",
-  "cpb_rights_modify": "Modifier",
-  "cpb_rights_add": "Ajouter",
-  "cpb_rights_table_rights": "Droits",
-  "cpb_rights_error": "Oups! nous n'arrivons pas a récupérer les informations sur les droits.",
-  "cpb_rights_add_error": "Oups! Nous n'arrivons pas à ajouter le droit.",
-  "cpb_rights_remove_error": "Oups! Nous n'arrivons pas à supprimer le droit.",
-  "cpb_rights_table_rights_value_readOnly": "Lecture seule",
-  "cpb_rights_table_rights_value_readWrite": "Lecture/Ecriture",
-  "cpb_rights_table_rights_add_success": "Le droit a bien été ajouté.",
-  "cpb_rights_table_rights_remove_success": "Le droit a bien été supprimé.",
-  "cpb_rights_add_title": "Ajouter un contact",
-  "cpb_rights_delete_title": "Supprimer un contact",
-  "cpb_rights_delete_question": "Voulez-vous vraiment supprimer les droits de {{nickname}}?",
-  "common_add": "Ajouter",
-  "delete": "Supprimer",
-  "common_cancel": "Annuler",
-  "common_confirm": "Confirmer",
-  "cpb_rights_note": "NB : Les contacts en lecture/écriture disposent de droits (modifications de l'infrastructure, utilisation du stockage) pouvant engendrer des coûts supplémentaires. Les contacts en lecture seule ne peuvent mener aucune action sur le projet. Le moyen de paiement par défaut du contact facturation est utilisé pour chacune des factures.",
-  "cpb_rights_table_nichandle": "Identifiant client",
-  "cpb_rights_table_email": "Email",
-  "pci_projects_project_contacts_title": "Contacts & droits"
-}
