--- packages/manager/apps/pci-project/public/translations/contacts/Messages_it_IT.json
@@ -1,26 +1,0 @@
-{
-  "cpb_rights_expl2": "Il contatto amministratore può aggiungere contatti in lettura/scrittura o in sola lettura per questo progetto.",
-  "cpb_rights_owner": "Contatto amministratore:",
-  "cpb_rights_billing_contact": "Contatto fatturazione:",
-  "cpb_rights_modify": "Modifica",
-  "cpb_rights_add": "Aggiungi",
-  "cpb_rights_table_rights": "Diritti",
-  "cpb_rights_error": "Ops! Impossibile recuperare le informazioni relative ai permessi.",
-  "cpb_rights_add_error": "Ops! Impossibile aggiungere il permesso.",
-  "cpb_rights_remove_error": "Ops! Impossibile rimuovere il permesso.",
-  "cpb_rights_table_rights_value_readOnly": "Sola lettura",
-  "cpb_rights_table_rights_value_readWrite": "Lettura/Scrittura",
-  "cpb_rights_table_rights_add_success": "Il permesso è stato aggiunto correttamente.",
-  "cpb_rights_table_rights_remove_success": "Il permesso è stato rimosso correttamente.",
-  "cpb_rights_add_title": "Aggiungi un contatto",
-  "cpb_rights_delete_title": "Elimina un contatto",
-  "cpb_rights_delete_question": "Vuoi davvero eliminare i permessi di {{nickname}}?",
-  "common_add": "Aggiungi",
-  "delete": "Elimina",
-  "common_cancel": "Annulla",
-  "common_confirm": "Conferma",
-  "cpb_rights_note": "Attenzione! Alcuni permessi assegnati ai contatti in lettura/scrittura (modifica dell'infrastruttura, utilizzo dello storage) possono generare costi aggiuntivi. I contatti in sola lettura non possono eseguire operazioni sul progetto. La modalità di pagamento predefinita del contatto di fatturazione viene utilizzata per il saldo di tutte le fatture.",
-  "cpb_rights_table_nichandle": "Identificativo cliente",
-  "cpb_rights_table_email": "Email",
-  "pci_projects_project_contacts_title": "Contatti e diritti"
-}
