--- packages/manager/apps/pci-project/public/translations/contacts/Messages_pl_PL.json
@@ -1,26 +1,0 @@
-{
-  "cpb_rights_expl2": "Kontakt administracyjny może również dodawać kontakty do zapisu/odczytu lub do samego odczytu dla tego projektu.",
-  "cpb_rights_owner": "Kontakt administracyjny:",
-  "cpb_rights_billing_contact": "Kontakt księgowy:",
-  "cpb_rights_modify": "Zmień",
-  "cpb_rights_add": "Dodaj",
-  "cpb_rights_table_rights": "Uprawnienia",
-  "cpb_rights_error": "Wystąpił problem z pobraniem informacji na temat uprawnień.",
-  "cpb_rights_add_error": "Wystąpił problem z przyznaniem uprawnień.",
-  "cpb_rights_remove_error": "Wystąpił problem z usunięciem uprawnień.",
-  "cpb_rights_table_rights_value_readOnly": "Do odczytu",
-  "cpb_rights_table_rights_value_readWrite": "Odczyt/Zapis",
-  "cpb_rights_table_rights_add_success": "Uprawnienia zostały przyznane.",
-  "cpb_rights_table_rights_remove_success": "Uprawnienia zostały usunięte.",
-  "cpb_rights_add_title": "Dodaj kontakt",
-  "cpb_rights_delete_title": "Usuń kontakt",
-  "cpb_rights_delete_question": "Czy na pewno chcesz usunąć uprawnienia dla {{nickname}}?",
-  "common_add": "Dodaj",
-  "delete": "Usuń",
-  "common_cancel": "Anuluj",
-  "common_confirm": "Potwierdź",
-  "cpb_rights_note": "Informacja: Kontakty z uprawnieniami do odczytu/zapisu dysponują uprawnieniami (zmiany infrastruktury, korzystanie z przestrzeni dyskowej), które mogą generować dodatkowe koszty. Kontakty z uprawnieniami tylko do odczytu nie mogą wykonywać operacji w ramach projektu. Domyślny sposób płatności skonfigurowany przez kontakt księgowy jest używany do regulowania wszystkich faktur.",
-  "cpb_rights_table_nichandle": "Identyfikator klienta",
-  "cpb_rights_table_email": "E-mail",
-  "pci_projects_project_contacts_title": "Kontakty i uprawnienia"
-}
