--- packages/manager/apps/pci-project/public/translations/contacts/Messages_pt_PT.json
@@ -1,26 +1,0 @@
-{
-  "cpb_rights_expl2": "Para este projeto, o admin pode adicionar contactos com permissões de leitura/escrita ou apenas de leitura.",
-  "cpb_rights_owner": "Contacto admin:",
-  "cpb_rights_billing_contact": "Contacto de faturação:",
-  "cpb_rights_modify": "Alterar",
-  "cpb_rights_add": "Adicionar",
-  "cpb_rights_table_rights": "Permissões",
-  "cpb_rights_error": "Oops! Não foi possível aceder à informação sobre as permissões.",
-  "cpb_rights_add_error": "Oops! Não foi possível eliminar a permissão.",
-  "cpb_rights_remove_error": "Oops! Não foi possível eliminar a permissão.",
-  "cpb_rights_table_rights_value_readOnly": "Apenas leitura",
-  "cpb_rights_table_rights_value_readWrite": "Leitura/Escrita",
-  "cpb_rights_table_rights_add_success": "A permissão foi adicionada com sucesso.",
-  "cpb_rights_table_rights_remove_success": "A permissão foi eliminada com sucesso.",
-  "cpb_rights_add_title": "Adicionar contacto",
-  "cpb_rights_delete_title": "Eliminar um contacto",
-  "cpb_rights_delete_question": "Quer mesmo eliminar as permissões de {{nickname}}?",
-  "common_add": "Adicionar",
-  "delete": "Eliminar",
-  "common_cancel": "Anular",
-  "common_confirm": "Confirmar",
-  "cpb_rights_note": "NB: os contactos com acesso de leitura/escrita dispõem de permissões (alterações na infraestrutura, uso do espaço de armazenamento) que podem gerar custos adicionais. Os contactos só com acesso leitura não podem efetuar nenhuma ação no projeto. O meio de pagamento predefinido pelo contacto de faturação será usado em todas as faturas.",
-  "cpb_rights_table_nichandle": "Identificador de cliente",
-  "cpb_rights_table_email": "E-mail",
-  "pci_projects_project_contacts_title": "Contactos e direitos"
-}
